Refactor correlate function
This change updates the correlate function to proper multiregion
support. The logic behind the creation of `LocationJson` was seriously
reworked. Most intermidiate representations now include the region.

Correlate now takes the region from the telegram, corresponding cli
argument is removed. If region meta information is not avaliable, fields
of the RegionMetaInformation are set to null. In view of that any meta
information input for `correlate` subcommand is removed.

epsg3857 is now calculated via a public method, all corresponding logic
is moved out to `tlms::locations`

Some minor changes to format strings to keep clippy happy